Fix(stock): Yahooファイナンスの株価取得処理を更新

+++ modules/stock/src/main/kotlin/com/example/stock/provider/StockInfo.kt
@@ -3,7 +3,7 @@ package com.example.stock.provider
 import java.time.LocalDate
 
 data class StockInfo(
-    val price: Int?,
+    val price: Double?,
     val dividend: Double?,
     val earningsDate: LocalDate?
 )
